Improve the comment for `go_sdk.host()` in the installation docs (#4230)
**What type of PR is this?**

Documentation

**What does this PR do?**

Explicitly call out `go_sdk.host()` as an alternative to
`go_sdk.download()` in the Bzlmod installation documentation, and point
clearly to the nota bene associated with it.

**What does this PR do? Why is it needed?**

Avoiding people from copying and pasting the code without understanding
that `go_sdk.host()` should be omitted when using `go_sdk.download()`
(as well as making sure that there's a clear pointer to its drawbacks).

**Which issues(s) does this PR fix?**

N/A