Add new setting //go/toolchain:sdk_name to allow the selection of a specific Go toolchain (#4242)
**What type of PR is this?**

Feature

**What does this PR do? Why is it needed?**

This commit adds a new setting
--@io_bazel_rules_go//go/toolchain:sdk_name which can be used to select
a registered Go toolchain via the name provided in go_download_sdk,
go_wrap_sdk, etc.

This fixes the problem of selecting a toolchain where the same Go
version is used, but different experiements are enabled or patches
applied.

**Which issues(s) does this PR fix?**

Fixes #4240

**Other notes for review**

Works also as a workaround for #3582