Update Issue template (#4258)
**What type of PR is this?**
Other

**What does this PR do? Why is it needed?**
It appears that the issue template is broken. The main reason for this
is a past change by GitHub, which now requires `ISSUE_TEMPLATE` to be a
folder instead of a file and uses yaml files instead of the original
markdown files.

Looks like PR template is working properly. 

ref:
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository

**Which issues(s) does this PR fix?**

Fixes #

**Other notes for review**
None