Mark gomock() hidden targets manual (#4265)
**What type of PR is this?**

Bug fix

**What does this PR do? Why is it needed?**

Imagine if you want to build mocks for interfaces that are only
available on certain operating systems/hardware platforms. In those
cases you simply want to add a "manual" tag to the gomock() target, and
only rely on it getting built if the resulting Go source file is part of
some go_library().

This change adds the "manual" tag to all of the hidden targets of this
rule, meaning that their build is skipped completely if the main
gomock() target is marked manual as well.

**Which issues(s) does this PR fix?**

One that requires me to carry a local patch around on the
github.com/buildbarn/bb-remote-execution side.

**Other notes for review**