Generate exports for stdlib (#4244)
**What type of PR is this?**

Feature

**What does this PR do? Why is it needed?**
This PR adds exports to the standard library which is useful for the
gopackagesdriver to avoid parsing files and various other tools that use
https://pkg.go.dev/golang.org/x/tools/go/packages.

**Which issues(s) does this PR fix?**
#3653 

**Other notes for review**
Basically a reimplementation of this PR:
https://github.com/bazel-contrib/rules_go/pull/3653 but with the comment
addressed about configuring it with `//go/config`. One additional thing
added is the exports files are part of the `GoSDK` provider so that
other custom rules can also gain access to the exported files.