Generate facts for targets not in nogo scope (#4268)
**What type of PR is this?**

Bug fix

**What does this PR do? Why is it needed?**

Nogo analyzers may need facts from their (transitive) dependencies to
operate correctly. We thus need to run nogo on all targets, even those
not in scope, but should only register the validation action for the
targets that are in scope to have them fail the build on nogo findings.

**Which issues(s) does this PR fix?**

Fixes #4241

**Other notes for review**