Stop using Label.workspace_name (#4282)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
4. Note that PRs updating dependencies and new Go versions are not
accepted.
   Please file an issue instead.
-->

**What type of PR is this?**

Other

**What does this PR do? Why is it needed?**

According to the Bazel documentation, it is deprecated. Label.repo_name
should be used instead.

https://bazel.build/rules/lib/builtins/Label#workspace_name

**Which issues(s) does this PR fix?**

I am working on an analysis tool for Bazel BUILD/*.bzl files. It is
currently unable to process Label.workspace_name. Adding support for it
is trivial, however I thought it would be a good idea to fix up existing
rules instead.

**Other notes for review**