Fail when expected files are not produced by protoc (#4287)
**What type of PR is this?**
Bug fix

**What does this PR do? Why is it needed?**
This PR fail the action if a plugin does not produce a valid Go file
that it is supposed to produce. This disallows people to add gRPC plugin
to a `go_proto_library` target that has no service definition, but
avoids the cache poisoning due to bugs in proto plugins. Because Gazelle
is able to switch between proto and grpc plugins depending on the
existence of service definition, existing repos managed by Gazelle will
not be affected.

**Which issues(s) does this PR fix?**

Fixes #3949

**Other notes for review**
Credits to @r-hang for coming up with the patch