fix the reporting message for getFixes (#4299)
**What type of PR is this?**
Improve the report message to avoid confusing developers.


> Bug fix


**What does this PR do? Why is it needed?**
Two issues:
- issue if there are no fixes at all, getFixes still produces "some
suggested fixes are invalid or have conflicts with other fixes:". The
right behavior is to not produce such message.
- if there are multiple suggested fixes, getFixes produces "some
suggested fixes are invalid or have conflicts with other fixes:". A
better solution is to, for each suggested fix, records how it fails
(invalid or conflicting).