Producing "fail" event when a test case panics (#4306)
**What type of PR is this?**
Bug fix

**What does this PR do? Why is it needed?**
When rules_go runs a test that panics, it doesn't produce the final
"FAIL" line, which the JSON converter relies on to flush the buffer at
the end. This PR writes the "FAIL" line to the JSON converter whenever a
test exits with a non-zero code to make sure the buffer is always
flushed.

**Which issues(s) does this PR fix?**

Fixes #4280

**Other notes for review**
When a test fails normally, the test would produce the "FAIL" line. When
this happens, the JSON converter would get two "FAIL" lines, and flush
the buffer twice at the end, which is fine.

If we run a test with `go test` and it panics, the "FAIL" line is still
produced. It is still a mystery why `bazel test` wouldn't produce the
"FAIL" line. As @jayconrod pointed out, the test to json logic in Go is
not well-designed. It may not worth fixing, but I am open to
suggestions.