fix(cgo): exclude crypto/internal/sysrand/internal/seccomp from SDK srcs (#4311)
**What type of PR is this?**

Bug fix

**What does this PR do? Why is it needed?**

The `crypto/internal/sysrand/internal/seccomp` package fails to compile
with linux kernel sources prior to 3.17. It's an internal package that
is only used by unit tests, so compiling it as part of the monolithic
SDK target is a waste of time anyway, so just exclude it.

**Which issues(s) does this PR fix?**

Fixes #4310