Add `go_sdk.from_file` to read the SDK version from `go.mod`. (#4305)
**What type of PR is this?**

Feature

**What does this PR do? Why is it needed?**

Adds a `go_sdk.from_file` rule to the extension that allows modules to
say that their SDK version should be read from their `go.mod` file. This
allows callers to avoid repeating themselves and stay within the Go tool
environment for version specification, allowing better integration with
other tools.

**Which issues(s) does this PR fix?**

Fixes #4292

**Other notes for review**

The toolchain version is selected based on the rules in
https://go.dev/doc/toolchain

This change copies parts of the `go_mod.bzl` parser from the Gazelle
repository:

https://github.com/bazel-contrib/bazel-gazelle/blob/master/internal/bzlmod/go_mod.bzl

A test based on the `go_download_sdk` test sets up a module-under-test
with a `go.mod` and
checks that the version specified in it is used to run the test in that
repo-under-test.