Export new location of support files needed for using Go's WASM output. (#4322)
**What type of PR is this?**


Bug fix

**What does this PR do? Why is it needed?**

In order to run code with GOOS=js GOARCH=wasm in the browser, you need
to copy a `.js` file from the toolchain.
This file location was changed in Go 1.24, which is now also exported in
this PR.

**Which issues(s) does this PR fix?**

n/a

**Other notes for review**

See also this info from the Go wiki:

> For Go 1.23 and earlier, the wasm support files needed in this article
are located in misc/wasm, and the path should be replaced when
performing operations with files such as lib/wasm/wasm_exec.js.

[source](https://go.dev/wiki/WebAssembly)